move services to constants.js and refracture codes
