+++ src/components/Calendar/Calendar.jsx
@@ -1,15 +1,15 @@
-import React from "react";
+import React, { useState } from "react";
 import "./Calendar.css";
-import { bookingCalendarOptions } from "../../utils/constants";
-("../../utils/constants");
+import { services } from "../../utils/constants";
 
-function Calendar() {
+function Calendar(selectedSrc) {
   //CalendarOptions to filter to certain calendar when different buttons are clicked
   return (
     <div className="calendar">
+      {/* insert bookingCalendarOptions */}
       <iframe
         className="calendar__content"
-        src="https://calendar.google.com/calendar/appointments/schedules/AcZssZ3l6mBnpsy_-h10nvyq9L98j15IoN1AkF7nLt0Q7wUtfbWomYRrt-fmKBEvTQS7_V_-XoK071kb?gv=true"
+        src={selectedSrc}
         // style={{ border: 0 }}
         // width="100%"
         // height="600"
